Move IP state from NIC to NetworkEndpoint/Protocol
* Add network address to network endpoints
Hold network-specific state in the NetworkEndpoint instead of the stack.
This results in the stack no longer needing to "know" about the network
endpoints and special case certain work for various endpoints
(e.g. IPv6 DAD).

* Provide NetworkEndpoints with an NetworkInterface interface
Instead of just passing the NIC ID of a NIC, pass an interface so the
network endpoint may query other information about the NIC such as
whether or not it is a loopback device.

* Move NDP code and state to the IPv6 package
NDP is IPv6 specific so there is no need for it to live in the stack.

Fixes: 3871
PiperOrigin-RevId: 326864960